optimization and stability

+++ skillcrawl.py
@@ -8,6 +8,13 @@ import time
 from urllib.parse import quote_plus
 from fuzzywuzzy import fuzz
 import pyfiglet
+import os
+import json
+
+CACHE_DIR = 'cache'
+CACHE_FILE = 'pdf_cache.json'
+
+os.makedirs(CACHE_DIR, exist_ok=True)
 
 skill_extractor = SkillExtractor()
 
@@ -16,8 +23,8 @@ def print_colored_text(text: str, color_code: str) -> None:
     print(f"\033[{color_code}m{text}\033[0m")
 
 def print_logo():
-    logo = pyfiglet.figlet_format("SKILLCRAWL")
-    print(logo)
+   logo = pyfiglet.figlet_format("SKILLCRAWL")
+   print(logo)
 
 def print_horizontal_line(length: int) -> None:
     print('=' * length)
@@ -43,6 +50,14 @@ def contains_no_lowercase_letters(text: str) -> bool:
 
 # Main Functions
 def download_pdf(url: str, save_path: str) -> str:
+    curriculum_folder = "curriculum"
+    os.makedirs(curriculum_folder, exist_ok=True)
+    save_path = os.path.join(curriculum_folder, save_path)
+
+    if os.path.exists(save_path):
+        print("Curriculum already retrieved!")
+        return save_path
+
     print(f"Downloading PDF from {url}")
     response = requests.get(url)
     with open(save_path, 'wb') as file:
@@ -50,6 +65,12 @@ def download_pdf(url: str, save_path: str) -> str:
     return save_path
 
 def extract_text_from_pdf(pdf_file_path: str) -> list:
+    cache_key = f"text_{pdf_file_path}"
+    if is_cached(cache_key):
+        print(f"Loading text from cache for {pdf_file_path}")
+        print_loading_line(25)
+        return load_cache()[cache_key]
+    
     print(f"Extracting text from PDF: {pdf_file_path}")
     print_loading_line(25)
     page_texts = []
@@ -57,23 +78,57 @@ def extract_text_from_pdf(pdf_file_path: str) -> list:
         for page in pdf.pages:
             text = page.extract_text() or ''
             page_texts.append(text)
+
+    cache = load_cache()
+    cache[cache_key] = page_texts
+    save_cache(cache)
     return page_texts
 
-def extract_text_after_marker(text: list, marker: str) -> str:
-    print_colored_text(f" >>> Extracting text after marker: {marker}", 34)
+
+def extract_text_after_marker(text: list, markers: list) -> str:
+    cache_key = f"text_after_marker_{hash(''.join(markers))}"
+
+    if is_cached(cache_key):
+        print(f"Loading text after markers from cache.")
+        print_loading_line(25)
+        return load_cache()[cache_key]
+
+    print_colored_text(f" >>> Extracting text after markers: {', '.join(markers)}", 34)
     full_text = '\n'.join(text)
-    marker_index = full_text.lower().find(marker.lower())
-    if marker_index == -1:
-        print("Marker not found in the text.")
-        return ""
-    return full_text[marker_index + len(marker):]
+    for marker in markers:
+        marker_index = full_text.lower().find(marker.lower())
+        if marker_index != -1:
+            print_colored_text(f"Found marker: {marker}", 32)
+            return full_text[marker_index + len(marker):]
+    
+    print("No marker found in the text.")
+    return ""
 
 def split_by_semester(text: str) -> list:
+    cache_key = f"semesters_{hash(text)}"
+    if is_cached(cache_key):
+        print(f"Loading semesters from cache")
+        return load_cache()[cache_key]
+    
     print("Splitting text by semester")
     semesters = re.split(r'(?i)(\b\d+\s*(?:st|nd|rd|th)?\s*Semester\b)', text)
-    combined_semesters = [''.join(semesters[i:i+2]) for i in range(1, len(semesters), 2)]
+    combined_semesters = []
+    seen_semesters = set()
+
+    for i in range(1, len(semesters), 2):
+        semester = ''.join(semesters[i:i+2])
+        if semester.lower() not in seen_semesters:
+            combined_semesters.append(semester)
+            seen_semesters.add(semester.lower())
+
     print(f"Found semesters: {len(combined_semesters)}")
+    
+    cache = load_cache()
+    cache[cache_key] = combined_semesters
+    save_cache(cache)
+    
     return combined_semesters
+
 
 def clean_lesson_name(name: str) -> str:
     return re.sub(r'\s*\(.*?\)\s*', '', name).strip()
@@ -113,26 +168,30 @@ def process_pages_by_lesson(pages: list) -> dict:
                     lesson_dict[lesson_name] = lesson_description
     return lesson_dict
 
+import requests
+
 def extract_and_get_title(skill_url):
     try:
+        if not skill_url.startswith("http://data.europa.eu/esco/skill/"):
+            print("Invalid skill URL format.")
+            return "Error: Invalid URL format"
         skill_id = skill_url.split('/skill/')[1]
-        base_url = "https://esco.ec.europa.eu/en/classification/skill?uri=http%3A%2F%2Fdata.europa.eu%2Fesco%2Fskill%2F"
-        combined_url = base_url + quote_plus(skill_id)
-        response = requests.get(combined_url)
-        
+
+        api_url = f"https://ec.europa.eu/esco/api/resource/skill?uri={skill_url}"
+        response = requests.get(api_url)
+
         if response.status_code == 200:
-            soup = BeautifulSoup(response.text, 'html.parser')
-            title = soup.title.string.strip() if soup.title else "No title found"
-            if '| ESCO' in title:
-                title = title.split(' | ESCO')[0].strip()
-            return title
+            data = response.json()
+            skill_title = data.get('preferredLabel', {}).get('en-us', None)
+            return skill_title
         else:
-            print(f"Failed to retrieve the page. Status code: {response.status_code}")
-            return "Error: Failed to fetch page"
+            return None
+
     except Exception as e:
         print(f"Error: {e}")
         return "Error: Exception occurred"
 
+    
 
 def display_menu():
     print_logo()
@@ -156,13 +215,24 @@ def display_menu():
     return command, lesson_name
 
 def get_skills_for_lesson(all_data, lesson_name, skills=False, skillname=False, threshold=80):
+    cache = load_cache()
     print_horizontal_line(50)
+    
+    cached_skills = cache.get(lesson_name.lower())
+    if cached_skills:
+        print_colored_text(f"Using cached skills for '{lesson_name}'", 33)
+        print_horizontal_small_line(50)
+        for skill_url in cached_skills:
+            print_colored_text(f'      Skill URL: {skill_url}', "32")
+        return
+
     for semester, lessons in all_data.items():
         matched_lessons = [(lesson, fuzz.partial_ratio(lesson_name.lower(), lesson.lower())) for lesson in lessons]
     
         matched_lessons = [lesson for lesson, score in matched_lessons if score >= threshold]
         
         if matched_lessons:
+            print_horizontal_line(50)
             print_colored_text(f'{semester}:', 33)
             print_horizontal_small_line(50)
             for lesson_name in matched_lessons:
@@ -184,6 +254,11 @@ def get_skills_for_lesson(all_data, lesson_name, skills=False, skillname=False, 
                         for skill_url in filtered_skills:
                             print_colored_text(f'      Skill URL: {skill_url}', "32")
                         print_green_line(50)
+
+
+                        cache[lesson_name.lower()] = list(filtered_skills)
+                        save_cache(cache)
+
                     else:
                         print_colored_text(f'    No skills found', "31")
                 elif skillname:
@@ -192,13 +267,19 @@ def get_skills_for_lesson(all_data, lesson_name, skills=False, skillname=False, 
                         print_green_line(50)
                         for skill_url in filtered_skills:
                             skill_name = extract_and_get_title(skill_url)
-                            print_colored_text(f'      [Skill]: {skill_name}', "32")
+                            if skill_name:
+                                print_colored_text(f'      [Skill]: {skill_name}', "32")
                         print_green_line(50)
+
+                        cache[lesson_name.lower()] = list(filtered_skills)
+                        save_cache(cache)
                     else:
                         print_colored_text(f'    No skills found!', "31")
                 print_horizontal_small_line(25)
         else:
             print(f"Lesson '{lesson_name}' not found in {semester}")
+    print_horizontal_line(25)
+
 
 def parse_args():
     """Parse command-line arguments."""
@@ -223,6 +304,24 @@ def parse_args():
                 lesson_name = ' '.join(sys.argv[2:])
     return simplified_mode, skills_mode, show_descr, skillname_mode, lesson_name
 
+def load_cache():
+    cache_file_path = os.path.join(CACHE_DIR, CACHE_FILE)
+    if os.path.exists(cache_file_path):
+        with open(cache_file_path, 'r') as cache_file:
+            return json.load(cache_file)
+    return {}
+
+# Save cache to file
+def save_cache(data):
+    cache_file_path = os.path.join(CACHE_DIR, CACHE_FILE)
+    with open(cache_file_path, 'w') as cache_file:
+        json.dump(data, cache_file, indent=4)
+
+# Check if data is already cached
+def is_cached(key):
+    cache = load_cache()
+    return key in cache
+
 def main(url: str, simplified: bool, skills: bool, show_descr: bool, skillname: bool, lesson_name: str = None):
     print_yellow_line(50)
     pdf_file_path = 'Program_of_Studies_2020-2021.pdf'
@@ -230,7 +329,7 @@ def main(url: str, simplified: bool, skills: bool, show_descr: bool, skillname: 
 
     pages = extract_text_from_pdf(pdf_file_path)
 
-    marker = 'Course Outlines'
+    marker = ['Course Outlines', 'Course Content']
     text_after_marker = extract_text_after_marker(pages, marker)
 
     semesters = split_by_semester(text_after_marker)
@@ -309,18 +408,43 @@ def main(url: str, simplified: bool, skills: bool, show_descr: bool, skillname: 
                         print_green_line(50)
                         for skill_url in filtered_skills:
                             skill_name = extract_and_get_title(skill_url)
-                            print_colored_text(f'      [Skill]: {skill_name}', "32")
+                            if skill_name:
+                                print_colored_text(f'      [Skill]: {skill_name}', "32")
                         print_green_line(50)
                     else:
                         print_colored_text(f'    No skills found!', "31")
                     print_horizontal_small_line(25)
 
-    if simplified:
-        print_horizontal_small_line(50)
-        print_horizontal_line(50)
+        elif simplified:
+            print_horizontal_small_line(50)
+            for semester, lessons in all_data.items():
+                print_horizontal_line(50)
+                print_colored_text(f'{semester}:', 33)
+                print_horizontal_line(50)
+                for lesson_name, lesson_description in lessons.items():
+                    print_colored_text(f'  {lesson_name}', "32")
+            print_horizontal_small_line(50)
+
+    command, lesson_name = display_menu()
+
+    if command == 'descr':
+        main(pdf_url, simplified=False, skills=False, show_descr=True, skillname=False)
+    elif command == 'skills':
+        main(pdf_url, simplified=False, skills=True, show_descr=False, skillname=False, lesson_name=lesson_name)
+    elif command == 'skillname':
+        main(pdf_url, simplified=False, skills=False, show_descr=False, skillname=True, lesson_name=lesson_name)
+    elif command == 'simplified':
+        main(pdf_url, simplified=True, skills=False, show_descr=False, skillname=False)
+    elif command == 'exit':
+        print("Exiting program...")
+        sys.exit(0)
+    else:
+        print("Invalid choice. Please try again.")
+        main(url, simplified, skills, show_descr, skillname, lesson_name)
 
 if __name__ == "__main__":
     pdf_url = "https://www.uom.gr/assets/site/public/nodes/4254/9477-Program_of_Studies_2020-2021.pdf"
+    
 
     simplified_mode, skills_mode, show_descr, skillname_mode, lesson_name = parse_args()
